requirements.txt
fastapi
uvicorn
motor
aioredis
python-dotenv
pydantic
pytest
pytest-asyncio